+++ Minigin/Transform.h
@@ -6,10 +6,10 @@ namespace GameEngine
 	class Transform final
 	{
 	public:
-		[[nodiscard]] glm::vec3 GetPosition() const { return m_Position; }
-		void SetPosition(float x, float y, float z);
-		void Translate(float x, float y);
-		void Translate(const glm::vec2& vec);
+		[[nodiscard]] glm::ivec3 GetPosition() const { return m_Position; }
+		void SetPosition(int x, int y, int z);
+		void Translate(int x, int y);
+		void Translate(const glm::ivec2& vec);
 		void SetRotation(float angle);
 		void SetScale(float xScale, float yScale);
